+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+### Fixed
+
+- Prevent crashing when trying to cache an unpicklable value.
+
 ## [2.0.7] - 2020-08-21
 
 ### Fixed